requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
Naked==0.1.31
pycryptodome==3.9.8
python-docx==0.8.10
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
urllib3==1.25.10
loguru